requirements.txt
pydantic
discord.py
asyncpg
sqlalchemy
redis[hiredis]
colorlog
psycopg2-binary
tenacity
pytz